+++ src/main/java/edu/biblioteca/bookflow/Repository/CursosRepository.java
@@ -1,13 +1,10 @@
 package edu.biblioteca.bookflow.Repository;
 
 import org.springframework.stereotype.Repository;
-
 import edu.biblioteca.bookflow.Model.Cursos;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface CursosRepository extends JpaRepository<Cursos, Integer>{
-
+public interface CursosRepository extends JpaRepository<Cursos, Long>{
     
 }
